Moved commands out into their own objects.
